UAT merge and hotfix

See merge request arkthusk/hr-and-payroll-system!34

+++ src/main/java/com/apap/HrPayrollSystem/Utility/Penggajian.java
@@ -6,20 +6,20 @@ import com.apap.HrPayrollSystem.Model.PegawaiOutsourcingModel;
 public class Penggajian {
 	GajiModel gaji = new GajiModel();
 	PegawaiOutsourcingModel pegawai = new PegawaiOutsourcingModel();
-	float gaji_bruto = 0; 
-	float bpjstk = 0;
-	float bpjsk = 0;
-	float total_bpjs = 0;
-	float jumlah_gaji = 0;
-	float ptkp = 0;
-	float pkp = 0; 
-	float pph21 = 0;
-	float total_potongan = 0;
-	float gaji_netto = 0;
+	int gaji_bruto = 0; 
+	int bpjstk = 0;
+	int bpjsk = 0;
+	int total_bpjs = 0;
+	int jumlah_gaji = 0;
+	int ptkp = 0;
+	int pkp = 0; 
+	int pph21 = 0;
+	int total_potongan = 0;
+	int gaji_netto = 0;
 	
 
-	public Penggajian(GajiModel gaji, PegawaiOutsourcingModel pegawai, float gaji_bruto, float bpjstk, float bpjsk,
-			float total_bpjs, float jumlah_gaji, float ptkp, float pkp, float pph21, float total_potongan, float gaji_netto) {
+	public Penggajian(GajiModel gaji, PegawaiOutsourcingModel pegawai, int gaji_bruto, int bpjstk, int bpjsk,
+			int total_bpjs, int jumlah_gaji, int ptkp, int pkp, int pph21, int total_potongan, int gaji_netto) {
 		super();
 		this.gaji = gaji;
 		this.pegawai = pegawai;
@@ -51,83 +51,83 @@ public class Penggajian {
 		this.pegawai = pegawai;
 	}
 
-	public float getGaji_bruto() {
+	public int getGaji_bruto() {
 		return gaji_bruto;
 	}
 
-	public void setGaji_bruto(float gaji_bruto) {
+	public void setGaji_bruto(int gaji_bruto) {
 		this.gaji_bruto = gaji_bruto;
 	}
 
-	public float getBpjstk() {
+	public int getBpjstk() {
 		return bpjstk;
 	}
 
-	public void setBpjstk(float bpjstk) {
+	public void setBpjstk(int bpjstk) {
 		this.bpjstk = bpjstk;
 	}
 
-	public float getBpjsk() {
+	public int getBpjsk() {
 		return bpjsk;
 	}
 
-	public void setBpjsk(float bpjsk) {
+	public void setBpjsk(int bpjsk) {
 		this.bpjsk = bpjsk;
 	}
 
-	public float getTotal_bpjs() {
+	public int getTotal_bpjs() {
 		return total_bpjs;
 	}
 
-	public void setTotal_bpjs(float total_bpjs) {
+	public void setTotal_bpjs(int total_bpjs) {
 		this.total_bpjs = total_bpjs;
 	}
 
-	public float getJumlah_gaji() {
+	public int getJumlah_gaji() {
 		return jumlah_gaji;
 	}
 
-	public void setJumlah_gaji(float jumlah_gaji) {
+	public void setJumlah_gaji(int jumlah_gaji) {
 		this.jumlah_gaji = jumlah_gaji;
 	}
 
-	public float getPtkp() {
+	public int getPtkp() {
 		return ptkp;
 	}
 
-	public void setPtkp(float ptkp) {
+	public void setPtkp(int ptkp) {
 		this.ptkp = ptkp;
 	}
 
-	public float getPkp() {
+	public int getPkp() {
 		return pkp;
 	}
 
-	public void setPkp(float pkp) {
+	public void setPkp(int pkp) {
 		this.pkp = pkp;
 	}
 
-	public float getPph21() {
+	public int getPph21() {
 		return pph21;
 	}
 
-	public void setPph21(float pph21) {
+	public void setPph21(int pph21) {
 		this.pph21 = pph21;
 	}
 
-	public float getTotal_potongan() {
+	public int getTotal_potongan() {
 		return total_potongan;
 	}
 
-	public void setTotal_potongan(float total_potongan) {
+	public void setTotal_potongan(int total_potongan) {
 		this.total_potongan = total_potongan;
 	}
 
-	public float getGaji_netto() {
+	public int getGaji_netto() {
 		return gaji_netto;
 	}
 
-	public void setGaji_netto(float gaji_netto) {
+	public void setGaji_netto(int gaji_netto) {
 		this.gaji_netto = gaji_netto;
 	}
